"Fix Login y Register, Fix NavBar, Fix flow"
