+++ client/src/index.css
@@ -2,7 +2,7 @@
 :root {
   --violeta-primario: #B270A2;
   --verde-primario: #7CE1CE;
-  --azul-complementario: #083AA9;
+  --azul-complementario: #355070;
   --negro-backgound: #181818;
   --negro-navBar-filters-btn: #101010;
 }
