+++ client/src/pages/Login/Login.jsx
@@ -2,18 +2,19 @@ import { useEffect, useState } from "react";
 import { sessionGoogle, singIn } from "../../firebase/auth";
 import { useForm } from "react-hook-form";
 import { useHistory } from "react-router-dom";
-
-import style from "./Login.module.css";
 import { Link } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { postCart } from "../../redux/actions";
+import { getUserCart, postCart } from "../../redux/actions";
 import Swal from "sweetalert2";
+import Loader from "../Home/GIF_aparecer_BooksNook.gif"
+import style from "./Login.module.css";
 
 export default function Login() {
   const [error, setError] = useState("");
   const history = useHistory();
   const dispatch = useDispatch()
-  const { user, lastRoute } = useSelector(state => state)
+  const { user } = useSelector(state => state)
+  const [loader, setLoader] = useState(false);
 
   const {
     register,
@@ -21,27 +22,56 @@ export default function Login() {
     handleSubmit,
   } = useForm();
 
+  //---------------- Pasar carrito de invitado a base de datos de usuario cuando inicia sesión ---------------
+
+  let repeatedIdArrayCart = [];
+  let uniqueIdArrayCart = [];
+  if (localStorage.length && localStorage.cart) {
+    repeatedIdArrayCart = localStorage.getItem("cart").split(",");
+    uniqueIdArrayCart = [...new Set(repeatedIdArrayCart)];
+  }
+
   useEffect(() => {
     if (user && user.uid) {
       dispatch(postCart({ userId: user.uid, bookId: [false], suma: true }))
-      const Toast = Swal.mixin({
-        background: "#19191a",
-        color: "#e1e1e1",
-        toast: true,
-        position: "top-end",
-        showConfirmButton: false,
-        timer: 2000,
-        timerProgressBar: true,
-      });
-      Toast.fire({
-        icon: "success",
-        title: `Has iniciado sesión con la cuenta: ${user.email}`,
-      });
-      setTimeout(() => {
-        history.push(lastRoute);
-      }, 1000);
+      if (uniqueIdArrayCart.length) {
+        Swal.fire({
+          title: 'Tienes productos en tu carrito de invitado',
+          width: 650,
+          text: "¿Quieres pasar estos productos a tu carrito de usuario?",
+          icon: 'warning',
+          iconColor: "#355070",
+          showCancelButton: true,
+          background: "#19191a",
+          color: "#e1e1e1",
+          confirmButtonColor: '#355070',
+          cancelButtonColor: '#B270A2',
+          confirmButtonText: '¡Si! Guardar carrito',
+          cancelButtonText: "Cancelar"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            dispatch(
+              postCart({ userId: user.uid, bookId: uniqueIdArrayCart, suma: true })
+            );
+            setLoader(true);
+            setTimeout(function () {
+              dispatch(getUserCart(user.uid));
+              localStorage.clear();
+            }, 1900);
+            setTimeout(() => {
+              history.goBack();
+            }, 2000);
+          } else {
+            history.goBack();
+          }
+        })
+      } else {
+        history.goBack();
+      }
     }
   }, [user])
+
+  //---------------- END Pasar carrito de invitado a base de datos de usuario cuando inicia sesión ---------------
 
   const handleGoogle = async () => {
     try {
@@ -62,50 +92,53 @@ export default function Login() {
 
   return (
     <div className={style.login_body}>
-      <div className={style.container}>
-        <h2 className={style.login}>Ingresar</h2>
-        <form onSubmit={handleSubmit(onSubmit)} className={style.form}>
-          <hr></hr>
-          <div className={style.container_input}>
-            <input
-              required
-              className={style.inputLogin}
-              type="text"
-              {...register("email", {
-                required: true,
-              })}
-            ></input>
-            <span></span>
-            <label className={style.labelText}>Correo</label>
-            {errors.email?.type === "required" && (
-              <p className={style.error}>Obligatorio</p>
-            )}
-          </div>
-          <div className={style.container_input}>
-            <input
-              required
-              className={style.inputLogin}
-              type="password"
-              {...register("password", {
-                required: true,
-              })}
-            ></input>
-            <span></span>
-            <label className={style.labelText}>Contraseña</label>
-            {errors.password?.type === "required" && (
-              <p className={style.error}>Obligatorio</p>
-            )}
-          </div>
-          <button className={style.button_form}>Iniciar sesión</button>
-          {error && <p className={style.ver}>{error}</p>}
-          <div className={style.google} onClick={handleGoogle}>
-            Iniciar sesión con google
-          </div>
-          <Link to="/register">
-            <p className={style.link}>¿No tienes una cuenta? Registrate</p>
-          </Link>
-        </form>
-      </div>
+      {!loader
+        ?
+        <div className={style.container}>
+          <h2 className={style.login}>Ingresar</h2>
+          <form onSubmit={handleSubmit(onSubmit)} className={style.form}>
+            <hr></hr>
+            <div className={style.container_input}>
+              <input
+                required
+                className={style.inputLogin}
+                type="text"
+                {...register("email", {
+                  required: true,
+                })}
+              ></input>
+              <span></span>
+              <label className={style.labelText}>Correo</label>
+              {errors.email?.type === "required" && (
+                <p className={style.error}>Obligatorio</p>
+              )}
+            </div>
+            <div className={style.container_input}>
+              <input
+                required
+                className={style.inputLogin}
+                type="password"
+                {...register("password", {
+                  required: true,
+                })}
+              ></input>
+              <span></span>
+              <label className={style.labelText}>Contraseña</label>
+              {errors.password?.type === "required" && (
+                <p className={style.error}>Obligatorio</p>
+              )}
+            </div>
+            <button className={style.button_form}>Iniciar sesión</button>
+            {error && <p className={style.ver}>{error}</p>}
+            <div className={style.google} onClick={handleGoogle}>
+              Iniciar sesión con google
+            </div>
+            <Link to="/register">
+              <p className={style.link}>¿No tienes una cuenta? Registrate</p>
+            </Link>
+          </form>
+        </div>
+        : <img className={style.loader} src={Loader} alt="Loader" />}
     </div>
   );
 }
