+++ client/src/pages/Register/Register.jsx
@@ -5,8 +5,9 @@ import { Link, useHistory } from "react-router-dom";
 import style from "./Register.module.css";
 import Swal from "sweetalert2";
 import { useDispatch, useSelector } from "react-redux";
-import { useEffect } from "react";
-import { postCart } from "../../redux/actions";
+import { useEffect, useState } from "react";
+import { getUserCart, postCart } from "../../redux/actions";
+import Loader from "../Home/GIF_aparecer_BooksNook.gif"
 
 export default function Register() {
   const history = useHistory();
@@ -17,29 +18,61 @@ export default function Register() {
   } = useForm();
   const { user, lastRoute } = useSelector(state => state)
   const dispatch = useDispatch()
+  const [loader, setLoader] = useState(false);
+
+  //---------------- Pasar carrito de invitado a base de datos de usuario cuando inicia sesión ---------------
+
+  let repeatedIdArrayCart = [];
+  let uniqueIdArrayCart = [];
+  if (localStorage.length && localStorage.cart) {
+    repeatedIdArrayCart = localStorage.getItem("cart").split(",");
+    uniqueIdArrayCart = [...new Set(repeatedIdArrayCart)];
+  }
 
   useEffect(() => {
     if (user && user.uid) {
       dispatch(postCart({ userId: user.uid, bookId: [false], suma: true }))
-
-      const Toast = Swal.mixin({
-        background: "#19191a",
-        color: "#e1e1e1",
-        toast: true,
-        position: "top-end",
-        showConfirmButton: false,
-        timer: 2000,
-        timerProgressBar: true,
-      });
-      Toast.fire({
-        icon: "success",
-        title: `Te has registrado con la cuenta: ${user.email}`,
-      });
-      setTimeout(() => {
-        history.push(lastRoute);
-      }, 1000);
+      if (uniqueIdArrayCart.length) {
+        Swal.fire({
+          title: 'Tienes productos en tu carrito de invitado',
+          width: 650,
+          text: "¿Quieres pasar estos productos a tu carrito de usuario?",
+          icon: 'warning',
+          iconColor: "#355070",
+          showCancelButton: true,
+          background: "#19191a",
+          color: "#e1e1e1",
+          confirmButtonColor: '#355070',
+          cancelButtonColor: '#B270A2',
+          confirmButtonText: '¡Si! Guardar carrito',
+          cancelButtonText: "Cancelar"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            dispatch(
+              postCart({ userId: user.uid, bookId: uniqueIdArrayCart, suma: true })
+            );
+            setLoader(true);
+            setTimeout(function () {
+              dispatch(getUserCart(user.uid));
+              localStorage.clear();
+            }, 1998);
+            setTimeout(() => {
+              setLoader(false);
+            }, 1999);
+            setTimeout(() => {
+              history.goBack();
+            }, 2000);
+          } else {
+            history.goBack();
+          }
+        })
+      } else {
+        history.goBack();
+      }
     }
   }, [user])
+
+  //---------------- END Pasar carrito de invitado a base de datos de usuario cuando inicia sesión ---------------
 
   const onSubmit = async (data) => {
     try {
@@ -59,73 +92,75 @@ export default function Register() {
 
   return (
     <div className={style.login_body}>
-      <div className={style.container}>
-        <h2 className={style.login}>Crear cuenta</h2>
-        <form onSubmit={handleSubmit(onSubmit)} className={style.form}>
-          <div className={style.container_input}>
-            <input
-              required
-              className={style.input}
-              // placeholder="Email"
-              type="text"
-              {...register("email", {
-                required: true,
-                pattern: /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/g,
-              })}
-            ></input>
-            <span></span>
-            <label className={style.labelText}>Correo</label>
-            {errors.email?.type === "required" && (
-              <p className={style.error}>Obligatorio</p>
-            )}
-            {errors.email?.type === "pattern" && (
-              <div className={style.error}>
-                <p>Ingresa un correo válido</p>
-                <i class="fa-solid fa-circle-exclamation"></i>
-              </div>
-            )}
-          </div>
-          <div className={style.container_input}>
-            <input
-              required
-              className={style.input}
-              // placeholder="Contraseña"
-              type="password"
-              {...register("password", {
-                required: true,
-                pattern: /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/,
-              })}
-            ></input>
-            <span></span>
-            <label className={style.labelText}>Contraseña</label>
-            {errors.password?.type === "required" && (
-              <p className={style.error}>Obligatorio</p>
-            )}
-            {errors.password?.type === "pattern" && (
-              <div className={style.error}>
-                <p >Mínimo 8 caracteres, (letras y números)</p>
-                <i class="fa-solid fa-circle-exclamation"></i>
-              </div>
-            )}
-          </div>
-          <div className={style.container_input}>
-            <input
-              required
-              className={style.input}
-              // placeholder="Nombre Completo"
-              type="text"
-              {...register("fullName", {
-                required: true,
-                maxLength: 20,
-              })}
-            ></input>
-            <span></span>
-            <label className={style.labelText}>Nombre completo</label>
-            {errors.fullName?.type === "required" && (
-              <p className={style.error}>Obligatorio</p>
-            )}
-          </div>
-          {/* <div className={style.inputContainer}>
+      {!loader
+        ?
+        <div className={style.container}>
+          <h2 className={style.login}>Crear cuenta</h2>
+          <form onSubmit={handleSubmit(onSubmit)} className={style.form}>
+            <div className={style.container_input}>
+              <input
+                required
+                className={style.input}
+                // placeholder="Email"
+                type="text"
+                {...register("email", {
+                  required: true,
+                  pattern: /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/g,
+                })}
+              ></input>
+              <span></span>
+              <label className={style.labelText}>Correo</label>
+              {errors.email?.type === "required" && (
+                <p className={style.error}>Obligatorio</p>
+              )}
+              {errors.email?.type === "pattern" && (
+                <div className={style.error}>
+                  <p>Ingresa un correo válido</p>
+                  <i class="fa-solid fa-circle-exclamation"></i>
+                </div>
+              )}
+            </div>
+            <div className={style.container_input}>
+              <input
+                required
+                className={style.input}
+                // placeholder="Contraseña"
+                type="password"
+                {...register("password", {
+                  required: true,
+                  pattern: /^(?=.*[A-Za-z])(?=.*\d)[A-Za-z\d]{8,}$/,
+                })}
+              ></input>
+              <span></span>
+              <label className={style.labelText}>Contraseña</label>
+              {errors.password?.type === "required" && (
+                <p className={style.error}>Obligatorio</p>
+              )}
+              {errors.password?.type === "pattern" && (
+                <div className={style.error}>
+                  <p >Mínimo 8 caracteres, (letras y números)</p>
+                  <i class="fa-solid fa-circle-exclamation"></i>
+                </div>
+              )}
+            </div>
+            <div className={style.container_input}>
+              <input
+                required
+                className={style.input}
+                // placeholder="Nombre Completo"
+                type="text"
+                {...register("fullName", {
+                  required: true,
+                  maxLength: 20,
+                })}
+              ></input>
+              <span></span>
+              <label className={style.labelText}>Nombre completo</label>
+              {errors.fullName?.type === "required" && (
+                <p className={style.error}>Obligatorio</p>
+              )}
+            </div>
+            {/* <div className={style.inputContainer}>
           <input
             className={style.input}
             placeholder="Provincia"
@@ -138,7 +173,7 @@ export default function Register() {
             <p className={style.error}>Obligatorio</p>
           )}
         </div> */}
-          {/* <div className={style.inputContainer}>
+            {/* <div className={style.inputContainer}>
           <input
             className={style.input}
             placeholder="Ciudad"
@@ -151,7 +186,7 @@ export default function Register() {
             <p className={style.error}>Obligatorio</p>
           )}
         </div> */}
-          {/* <div className={style.inputContainer}>
+            {/* <div className={style.inputContainer}>
           <input
             className={style.input}
             placeholder="Codigo Postal"
@@ -172,7 +207,7 @@ export default function Register() {
             <p className={style.error}>Maximo 5 numeros</p>
           )}
         </div> */}
-          {/* <div className={style.inputContainer}>
+            {/* <div className={style.inputContainer}>
           <input
             className={style.input}
             placeholder="Direccion"
@@ -186,17 +221,18 @@ export default function Register() {
             <p className={style.error}>Obligatorio</p>
           )}
         </div> */}
-          {/* <div className={style.register}> */}
-          <button className={style.button_form}>Crear cuenta</button>
-          <div className={style.google} onClick={handleSignInGoogle}>
-            Registrarse con google
-          </div>
-          <Link to="/login">
-            <p className={style.link}>¿Ya tienes una cuenta? Inicia sesión</p>
-          </Link>
-          {/* </div> */}
-        </form>
-      </div>
+            {/* <div className={style.register}> */}
+            <button className={style.button_form}>Crear cuenta</button>
+            <div className={style.google} onClick={handleSignInGoogle}>
+              Registrarse con google
+            </div>
+            <Link to="/login">
+              <p className={style.link}>¿Ya tienes una cuenta? Inicia sesión</p>
+            </Link>
+            {/* </div> */}
+          </form>
+        </div>
+        : <img className={style.loader} src={Loader} alt="Loader" />}
     </div>
   );
 }
